requirements.txt
dill
numpy
opencv-python
Pillow
matplotlib